Module arrow_buffer::bigint

Modules

div 🔒

N-digit division

Macros

define_as_primitive 🔒

derive_op 🔒

Structs

ParseI256Error

An opaque error similar to std::num::ParseIntError

i256

A signed 256-bit integer

Enums

DivRemError 🔒

Error returned by i256::DivRem

Functions

mulx 🔒

Performs an unsigned multiplication of a * b returning a tuple of (low, high) where low contains the lower 128-bits of the result and high the higher 128-bits

parse_impl 🔒

Parse s with any sign and leading 0s removed

split_array 🔒

Temporary workaround due to lack of stable const array slicing See https://github.com/rust-lang/rust/issues/90091