Module arrow_buffer::bigint::div

N-digit division

Implementation heavily inspired by uint

Structs

ArrayPlusOne 🔒

An array of N + 1 elements

Functions

add_assign 🔒

Perform a += b

binop_slice 🔒

Converts an overflowing binary operation on scalars to one on slices

bits 🔒

Return the least number of bits needed to represent the number

div_rem

Unsigned, little-endian, n-digit division with remainder

div_rem_knuth 🔒

Use Knuth Algorithm D to compute numerator / divisor returning the quotient and remainder

div_rem_small 🔒

Division of numerator by a u64 divisor

div_rem_word 🔒

Perform narrowing division of a u128 by a u64 divisor, returning the quotient and remainder

full_mul_u64 🔒

Widening multiplication of an N-digit array with a u64

full_shl 🔒

Widening left shift of an N-digit array by at most 63 bits

full_shr 🔒

Narrowing right shift of an (N+1)-digit array by at most 63 bits

shl_word 🔒

Left shift of an N-digit array by at most 63 bits

sub_assign 🔒

Perform a -= b