arrow_buffer

Module bit_util

Utils for working with bits

Functions

apply_bitwise_binary_op

Applies a bitwise operation relative to another bit-packed byte slice (right) in place

apply_bitwise_unary_op

Apply a bitwise operation to a mutable buffer, updating it in place.

ceil

Returns the ceil of value/divisor

get_bit

Returns whether bit at position i in data is set or not

get_bit_raw⚠

Returns whether bit at position i in data is set or not.

round_upto_multiple_of_64

Returns the nearest number that is >= than num and is a multiple of 64

round_upto_power_of_2

Returns the nearest multiple of factor that is >= than num. Here factor must be a power of 2.

set_bit

Sets bit at position i for data to 1

set_bit_raw⚠

Sets bit at position i for data

unset_bit

Sets bit at position i for data to 0

unset_bit_raw⚠

Sets bit at position i for data to 0