arrow_buffer::buffer::boolean

Struct BooleanBuffer

pub struct BooleanBuffer {
    buffer: Buffer,
    offset: usize,
    len: usize,
}

A slice-able Buffer containing bit-packed booleans

BooleanBuffers can be creating using BooleanBufferBuilder

See Also

NullBuffer for representing null values in Arrow arrays

Fields
buffer: Bufferoffset: usizelen: usize
Implementations

impl BooleanBuffer

pub fn new(buffer: Buffer, offset: usize, len: usize) -> Self

Create a new BooleanBuffer from a Buffer, an offset and length in bits

Panics

This method will panic if buffer is not large enough

pub fn new_set(length: usize) -> Self

Create a new BooleanBuffer of length where all values are true

pub fn new_unset(length: usize) -> Self

Create a new BooleanBuffer of length where all values are false

pub fn collect_bool<F: FnMut(usize) -> bool>(len: usize, f: F) -> Self

Invokes f with indexes 0..len collecting the boolean results into a new BooleanBuffer

pub fn count_set_bits(&self) -> usize

Returns the number of set bits in this buffer

pub fn bit_chunks(&self) -> BitChunks<'_>

Returns a BitChunks instance which can be used to iterate over this buffer’s bits in u64 chunks

pub fn offset(&self) -> usize

Returns the offset of this BooleanBuffer in bits

pub fn len(&self) -> usize

Returns the length of this BooleanBuffer in bits

pub fn is_empty(&self) -> bool

Returns true if this BooleanBuffer is empty

pub fn shrink_to_fit(&mut self)

Free up unused memory.

pub fn value(&self, idx: usize) -> bool

Returns the boolean value at index i.

Panics

Panics if i >= self.len()

pub unsafe fn value_unchecked(&self, i: usize) -> bool

Returns the boolean value at index i.

Safety

This doesn’t check bounds, the caller must ensure that index < self.len()

pub fn values(&self) -> &[u8]

Returns the packed values of this BooleanBuffer not including any offset

pub fn slice(&self, offset: usize, len: usize) -> Self

Slices this BooleanBuffer by the provided offset and length

pub fn sliced(&self) -> Buffer

Returns a Buffer containing the sliced contents of this BooleanBuffer

Equivalent to self.buffer.bit_slice(self.offset, self.len)

pub fn ptr_eq(&self, other: &Self) -> bool

Returns true if this BooleanBuffer is equal to other, using pointer comparisons to determine buffer equality. This is cheaper than PartialEq::eq but may return false when the arrays are logically equal

pub fn inner(&self) -> &Buffer

Returns the inner Buffer

pub fn into_inner(self) -> Buffer

Returns the inner Buffer, consuming self

pub fn iter(&self) -> BitIterator<'_>

Returns an iterator over the bits in this BooleanBuffer

pub fn set_indices(&self) -> BitIndexIterator<'_>

Returns an iterator over the set bit positions in this BooleanBuffer

pub fn set_slices(&self) -> BitSliceIterator<'_>

Returns a BitSliceIterator yielding contiguous ranges of set bits