Struct arrow_buffer::builder::BooleanBufferBuilder

pub struct BooleanBufferBuilder {
    buffer: MutableBuffer,
    len: usize,
}

Builder for BooleanBuffer

Fields
buffer: MutableBufferlen: usize
Implementations

impl BooleanBufferBuilder

pub fn new(capacity: usize) -> Self

Creates a new BooleanBufferBuilder

pub fn new_from_buffer(buffer: MutableBuffer, len: usize) -> Self

Creates a new BooleanBufferBuilder from MutableBuffer of len

pub fn len(&self) -> usize

Returns the length of the buffer

pub fn set_bit(&mut self, index: usize, v: bool)

Sets a bit in the buffer at index

pub fn get_bit(&self, index: usize) -> bool

Gets a bit in the buffer at index

pub fn is_empty(&self) -> bool

Returns true if empty

pub fn capacity(&self) -> usize

Returns the capacity of the buffer

pub fn advance(&mut self, additional: usize)

Advances the buffer by additional bits

pub fn truncate(&mut self, len: usize)

Truncates the builder to the given length

If len is greater than the buffer’s current length, this has no effect

pub fn reserve(&mut self, additional: usize)

Reserve space to at least additional new bits. Capacity will be >= self.len() + additional. New bytes are uninitialized and reading them is undefined behavior.

pub fn resize(&mut self, len: usize)

Resizes the buffer, either truncating its contents (with no change in capacity), or growing it (potentially reallocating it) and writing false in the newly available bits.

pub fn append(&mut self, v: bool)

Appends a boolean v into the buffer

pub fn append_n(&mut self, additional: usize, v: bool)

Appends n additional bits of value v into the buffer

pub fn append_slice(&mut self, slice: &[bool])

Appends a slice of booleans into the buffer

pub fn append_packed_range(&mut self, range: Range<usize>, to_set: &[u8])

Append range bits from to_set

to_set is a slice of bits packed LSB-first into [u8]

Panics

Panics if to_set does not contain ceil(range.end / 8) bytes

pub fn as_slice(&self) -> &[u8]

Returns the packed bits

pub fn as_slice_mut(&mut self) -> &mut [u8]

Returns the packed bits

pub fn finish(&mut self) -> BooleanBuffer

Creates a BooleanBuffer

pub fn finish_cloned(&self) -> BooleanBuffer

Builds the BooleanBuffer without resetting the builder.