arrow_buffer::builder
Struct NullBufferBuilder

pub struct NullBufferBuilder {
    bitmap_builder: Option<BooleanBufferBuilder>,
    len: usize,
    capacity: usize,
}

Builder for creating the null bit buffer.

This builder only materializes the buffer when we append false. If you only append trues to the builder, what you get will be None when calling finish. This optimization is very important for the performance.

Fields
bitmap_builder: Option<BooleanBufferBuilder>len: usize

Store the length of the buffer before materializing.

capacity: usize
Implementations

impl NullBufferBuilder

pub fn new(capacity: usize) -> Self

Creates a new empty builder. capacity is the number of bits in the null buffer.

pub fn new_with_len(len: usize) -> Self

Creates a new builder with given length.

pub fn new_from_buffer(buffer: MutableBuffer, len: usize) -> Self

Creates a new builder from a MutableBuffer.

pub fn append_n_non_nulls(&mut self, n: usize)

Appends n trues into the builder to indicate that these n items are not nulls.

pub fn append_non_null(&mut self)

Appends a true into the builder to indicate that this item is not null.

pub fn append_n_nulls(&mut self, n: usize)

Appends n falses into the builder to indicate that these n items are nulls.

pub fn append_null(&mut self)

Appends a false into the builder to indicate that this item is null.

pub fn append(&mut self, not_null: bool)

Appends a boolean value into the builder.

pub fn append_slice(&mut self, slice: &[bool])

Appends a boolean slice into the builder to indicate the validations of these items.

pub fn finish(&mut self) -> Option<NullBuffer>

Builds the null buffer and resets the builder. Returns None if the builder only contains trues.

pub fn finish_cloned(&self) -> Option<NullBuffer>

Builds the NullBuffer without resetting the builder.

pub fn as_slice(&self) -> Option<&[u8]>

Returns the inner bitmap builder as slice

fn materialize_if_needed(&mut self)

fn materialize(&mut self)

pub fn as_slice_mut(&mut self) -> Option<&mut [u8]>

Return a mutable reference to the inner bitmap slice.

pub fn allocated_size(&self) -> usize

Return the allocated size of this builder, in bytes, useful for memory accounting.

impl NullBufferBuilder

pub fn len(&self) -> usize

Return the number of bits in the buffer.

pub fn is_empty(&self) -> bool

Check if the builder is empty.