arrow_buffer

Module pool

This module contains traits for memory pool traits and an implementation for tracking memory usage.

The basic traits are MemoryPool and MemoryReservation. And default implementation of MemoryPool is TrackingMemoryPool. Their relationship is as follows:

    (pool tracker)                        (resizable)           
 â”Œâ”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â” fn reserve() â”Œâ”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”
 â”‚ trait MemoryPool â”‚â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â–ºâ”‚ trait MemoryReservation â”‚
 â””â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”˜              â””â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”˜

Structs

Tracker ğŸ”’

TrackingMemoryPool

A simple MemoryPool that reports the total memory usage

Traits

MemoryPool

A pool of memory that can be reserved and released.

MemoryReservation

A memory reservation within a MemoryPool that is freed on drop