arrow_buffer::util

Module bit_mask

Utils for working with packed bit masks

Functions

or_write_u64_bytes 🔒 ⚠

Similar to write_u64_bytes, but this method ORs the offset addressed data and chunk instead of overwriting

read_bytes_to_u64 🔒 ⚠

Safety

set_bits

Util function to set bits in a slice of bytes.

set_upto_64bits 🔒 ⚠

Similar to set_bits but sets only upto 64 bits, actual number of bits set may vary. Returns a pair of the number of 0 bits and the number of bits set

write_u64_bytes 🔒 ⚠

Safety