arrow_buffer::util

Module bit_util

Utils for working with bits

Structs

U64UnalignedSlice 🔒

Centralized structure to handle a mutable u8 slice as a mutable u64 pointer.

Functions

align_to_byte 🔒

Align to byte boundary by applying operation to bits before the next byte boundary.

apply_bitwise_binary_op

Applies a bitwise operation relative to another bit-packed byte slice (right) in place

apply_bitwise_unary_op

Apply a bitwise operation to a mutable buffer, updating it in place.

byte_aligned_bitwise_bin_op_helper 🔒

Perform bitwise binary operation on byte-aligned buffers (i.e. not offsetting into a middle of a byte).

byte_aligned_bitwise_unary_op_helper 🔒

Perform bitwise unary operation on byte-aligned buffer.

ceil

Returns the ceil of value/divisor

get_bit

Returns whether bit at position i in data is set or not

get_bit_raw⚠

Returns whether bit at position i in data is set or not.

get_remainder_bits 🔒

Read remainder bits from a slice.

handle_mutable_buffer_remainder 🔒

Handle remainder bits (< 64 bits) for binary operations.

handle_mutable_buffer_remainder_unary 🔒

Handle remainder bits (< 64 bits) for unary operations.

read_u64 🔒

Read a u64 from a byte slice, padding with zeros if necessary

read_up_to_byte_from_offset 🔒

Read up to 8 bits from a byte slice starting at a given bit offset.

round_upto_multiple_of_64

Returns the nearest number that is >= than num and is a multiple of 64

round_upto_power_of_2

Returns the nearest multiple of factor that is >= than num. Here factor must be a power of 2.

set_bit

Sets bit at position i for data to 1

set_bit_raw⚠

Sets bit at position i for data

set_remainder_bits 🔒

Write remainder bits back to buffer while preserving bits outside the range.

unset_bit

Sets bit at position i for data to 0

unset_bit_raw⚠

Sets bit at position i for data to 0