Crate arrow_buffer

Low-level buffer abstractions for Apache Arrow Rust

Re-exports

pub use buffer::*;

pub use builder::*;

Modules

alloc

Defines the low-level Allocation API for shared memory regions

arith 🔒

bigint 🔒

bit_chunk_iterator

Types for iterating over bitmasks in 64-bit chunks

bit_iterator

Types for iterating over packed bitmasks

bit_mask

Utils for working with packed bit masks

bit_util

Utils for working with bits

buffer

Types of shared memory region

builder

Buffer builders

bytes 🔒

This module contains an implementation of a contiguous immutable memory region that knows how to de-allocate itself, Bytes. Note that this is a low-level functionality of this crate.

interval 🔒

native 🔒

util 🔒

Structs

IntervalDayTime

Value of an IntervalDayTime array

IntervalMonthDayNano

Value of an IntervalMonthDayNano array

i256

A signed 256-bit integer

Traits

ArrowNativeType

Trait expressing a Rust type that has the same in-memory representation as Arrow.

ToByteSlice

Allows conversion from supported Arrow types to a byte slice.