Crate arrow_buffer

Buffer abstractions for Apache Arrow

Modules

alloc

Defines memory-related functions, such as allocate/deallocate/reallocate memory regions, cache and allocation alignments.

bit_chunk_iterator

Types for iterating over bitmasks in 64-bit chunks

bit_util

Utils for working with bits

buffer

This module contains two main structs: Buffer and MutableBuffer. A buffer represents a contiguous memory region that can be shared via offsets.

Structs

Buffer

Buffer represents a contiguous memory region that can be shared with other buffers and across thread boundaries.

MutableBuffer

A MutableBuffer is Arrow’s interface to build a Buffer out of items or slices of items. Buffers created from MutableBuffer (via into) are guaranteed to have its pointer aligned along cache lines and in multiple of 64 bytes. Use MutableBuffer::push to insert an item, MutableBuffer::extend_from_slice to insert many items, and into to convert it to Buffer.

i256

A signed 256-bit integer

Traits

ArrowNativeType

Trait expressing a Rust type that has the same in-memory representation as Arrow. This includes i16, f32, but excludes bool (which in arrow is represented in bits).

ToByteSlice

Allows conversion from supported Arrow types to a byte slice.